+++ test-utils/test-utils-core/src/main/java/uk/gov/justice/services/test/utils/core/random/DateGenerator.java
@@ -1,0 +1,6 @@
+package uk.gov.justice.services.test.utils.core.random;
+
+public abstract class DateGenerator<T> extends Generator<T> {
+
+    public enum Direction {PAST, FUTURE}
+}